+++ ACP3/Core/DependencyInjection/ServiceContainerBuilder.php
@@ -16,7 +16,6 @@ use ACP3\Core\Modules;
 use ACP3\Core\Validation\DependencyInjection\RegisterValidationRulesPass;
 use ACP3\Core\View\Renderer\Smarty\DependencyInjection\RegisterSmartyPluginsPass;
 use ACP3\Core\WYSIWYG\DependencyInjection\RegisterWysiwygEditorsCompilerPass;
-use Psr\Log\LoggerInterface;
 use Symfony\Component\Config\FileLocator;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Loader\YamlFileLoader;
@@ -25,10 +24,6 @@ use Symfony\Component\HttpFoundation\Request as SymfonyRequest;
 
 class ServiceContainerBuilder extends ContainerBuilder
 {
-    /**
-     * @var LoggerInterface
-     */
-    private $logger;
     /**
      * @var ApplicationPath
      */
@@ -45,20 +40,15 @@ class ServiceContainerBuilder extends ContainerBuilder
     /**
      * ServiceContainerBuilder constructor.
      *
-     * @param LoggerInterface $logger
      * @param ApplicationPath $applicationPath
      * @param SymfonyRequest  $symfonyRequest
      * @param string          $applicationMode
      */
     public function __construct(
-        LoggerInterface $logger,
-        ApplicationPath $applicationPath,
-        SymfonyRequest $symfonyRequest,
-        $applicationMode
+        ApplicationPath $applicationPath, SymfonyRequest $symfonyRequest, $applicationMode
     ) {
         parent::__construct();
 
-        $this->logger = $logger;
         $this->applicationPath = $applicationPath;
         $this->symfonyRequest = $symfonyRequest;
         $this->applicationMode = $applicationMode;
@@ -68,7 +58,6 @@ class ServiceContainerBuilder extends ContainerBuilder
 
     private function setUpContainer()
     {
-        $this->set('core.logger.system_logger', $this->logger);
         $this->set('core.http.symfony_request', $this->symfonyRequest);
         $this->set('core.environment.application_path', $this->applicationPath);
         $this->setParameter('core.environment', $this->applicationMode);
@@ -112,7 +101,6 @@ class ServiceContainerBuilder extends ContainerBuilder
     }
 
     /**
-     * @param LoggerInterface                        $logger
      * @param \ACP3\Core\Environment\ApplicationPath $applicationPath
      * @param SymfonyRequest                         $symfonyRequest
      * @param string                                 $applicationMode
@@ -120,12 +108,9 @@ class ServiceContainerBuilder extends ContainerBuilder
      * @return ContainerBuilder
      */
     public static function create(
-        LoggerInterface $logger,
-        ApplicationPath $applicationPath,
-        SymfonyRequest $symfonyRequest,
-        $applicationMode
+        ApplicationPath $applicationPath, SymfonyRequest $symfonyRequest, $applicationMode
     ) {
-        return new static($logger, $applicationPath, $symfonyRequest, $applicationMode);
+        return new static($applicationPath, $symfonyRequest, $applicationMode);
     }
 
     /**