+++ ACP3/Core/Modules/ModuleDependenciesTrait.php
@@ -16,7 +16,7 @@ trait ModuleDependenciesTrait
      *
      * @return array
      */
-    protected function getModuleDependencies($path)
+    protected function getModuleDependencies(string $path): array
     {
         $dependencies = $this->getXml()->parseXmlFile($path, '/module/info/dependencies');